More refreshing optimisations
Widgets no longer refresh upon a rename or click
